fix(design-system): move DS component stories to stories/ folder
Storybook main.ts now globs ../stories/** (not ../src/**), so button and
icon stories were invisible. Move them to stories/DS-Button.stories.tsx
and stories/DS-Icon.stories.tsx with updated import paths.
